fix : 장바구니, 주문에 토큰 적용

+++ src/main/java/org/example/outlivryteamproject/domain/order/service/OrderServiceImpl.java
@@ -9,6 +9,7 @@ import org.example.outlivryteamproject.domain.order.entity.Order;
 import org.example.outlivryteamproject.domain.order.repository.OrderRepository;
 import org.example.outlivryteamproject.domain.store.repository.StoreRepository;
 import org.example.outlivryteamproject.domain.user.entity.User;
+import org.example.outlivryteamproject.domain.user.enums.UserRole;
 import org.example.outlivryteamproject.domain.user.repository.UserRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -48,36 +49,64 @@ public class OrderServiceImpl implements OrderService{
     }
 
     @Override
-    public OrderResponseDto findByOrderId(Long orderId) {
+    public OrderResponseDto findByOrderId(Long userId, Long orderId) {
 
         Order findedOrder = orderRepository.findByOrderIdOrElseThrow(orderId);
+
+        if (!findedOrder.getUser().getId().equals(userId)) {
+            throw new IllegalArgumentException("접근할 수 없습니다");
+        }
 
         return new OrderResponseDto(findedOrder);
     }
 
     @Override
     @Transactional
-    public void receivedOrder(Long orderId) {
+    public void receivedOrder(Long userId, Long orderId) {
 
+        User findedUser = userRepository.findByIdOrElseThrow(userId);
         Order findedOrder = orderRepository.findByOrderIdOrElseThrow(orderId);
+
+        if (findedUser.getUserRole().equals(UserRole.USER)) {
+            throw new IllegalArgumentException("권한이 없습니다.");
+        }
+
+        if (!findedOrder.getStore().getUser().equals(findedUser)) {
+            throw new IllegalArgumentException("사장님만 변경할 수 있습니다.");
+        }
+
         findedOrder.changeReceived();
-
     }
 
     @Override
     @Transactional
-    public void deliveryOrder(Long orderId) {
+    public void deliveryOrder(Long userId, Long orderId) {
 
+        User findedUser = userRepository.findByIdOrElseThrow(userId);
         Order findedOrder = orderRepository.findByOrderIdOrElseThrow(orderId);
+
+        if (findedUser.getUserRole().equals(UserRole.USER)) {
+            throw new IllegalArgumentException("권한이 없습니다.");
+        }
+
+        if (!findedOrder.getStore().getUser().equals(findedUser)) {
+            throw new IllegalArgumentException("사장님만 변경할 수 있습니다.");
+        }
+
         findedOrder.changeDelivery();
-
     }
 
     @Override
     @Transactional
-    public void deleteOrder(Long orderId) {
+    public void deleteOrder(Long userId, Long orderId) {
 
+        User findedUser = userRepository.findByIdOrElseThrow(userId);
         Order findedOrder = orderRepository.findByOrderIdOrElseThrow(orderId);
+
+        if (!findedUser.equals(findedOrder.getUser())) {
+            throw new IllegalArgumentException("작성자만 삭제할 수 있습니다.");
+        }
+
         orderRepository.delete(findedOrder);
 
     }
